+++ frontend/app/src/pages/dailyreport/index.tsx
@@ -11,30 +11,41 @@ import { Flex } from '../../components/layout/Flex'
 import { DailyReportsList } from '../../components/organisms/DailyReport'
 import { Sidebar } from '../../components/organisms/Sidebar'
 import Layout from '../../components/templates/Layout'
+import { NextPage } from 'next/types'
+import { DailyReportType } from '../../types/types'
+import { type } from 'os'
 
-const dailyReport = () => {
-  const Today = new Date()
-  const [dailyReports, setDailyReports] = useState([])
-  const [userId, setUserId] = useState(false)
-  const [loading, setLoading] = useState(true)
-  const [reportDateOn, setReportDateOn] = useState(Today)
-  const [editReportDateOn, setEditReportDateOn] = useState(Today)
-  const [inputData, setInputData] = useState({
+const dailyReport: NextPage = () => {
+  const Today: Date = new Date()
+  const [dailyReports, setDailyReports] = useState<DailyReportType[]>([])
+  const [userId, setUserId] = useState<number | boolean>(false)
+  const [loading, setLoading] = useState<boolean>(true)
+  const [reportDateOn, setReportDateOn] = useState<Date>(Today)
+  const [editReportDateOn, setEditReportDateOn] = useState<Date>(Today)
+  const [inputData, setInputData] = useState<{
+    text: string
+    time: number | string
+  }>({
     text: '',
     time: '',
   })
-  const [editedDailyReport, setEditedDailyReport] = useState(false)
+  const [editedDailyReport, setEditedDailyReport] = useState<DailyReportType>({
+    id: NaN,
+    reportDateOn: '',
+    text: '',
+    time: NaN,
+  })
   const { currentUserId, combinedTime, setCombinedTime } =
     useContext(AuthContext)
-  const editInputRef = useRef(null)
+  const editInputRef = useRef<HTMLInputElement>(null)
 
   useEffect(() => {
-    setUserId(currentUserId)
-    console.log(currentUserId, 'current')
+    if (typeof currentUserId === 'number') {
+      setUserId(currentUserId)
+    }
   }, [currentUserId])
 
   useEffect(() => {
-    console.log(typeof userId, 'userId')
     if (typeof userId == 'number') {
       getDailyReportsList(userId).then((dailyReportsData) => {
         setDailyReports(dailyReportsData)
@@ -44,28 +55,29 @@ const dailyReport = () => {
   }, [userId])
 
   useEffect(() => {
-    if (editedDailyReport) {
-      editInputRef.current.focus()
-      console.log(editInputRef.current)
-    }
+    editInputRef.current?.focus()
   }, [editedDailyReport.id])
 
   useEffect(() => {
     if (typeof userId == 'number') {
       getUserProfileData(userId).then((data) => {
-        setCombinedTime(data.combinedTime)
+        if (typeof data.combinedTime === 'number') {
+          setCombinedTime(data.combinedTime)
+        }
       })
     }
   }, [dailyReports])
 
-  const onChangeInputData = (key, value) => {
+  const onChangeInputData = (key: string, value: string | number) => {
     setInputData({ ...inputData, [key]: value })
   }
 
-  const onChangeEditInput = (key, value) => {
-    setEditedDailyReport({ ...editedDailyReport, [key]: value })
+  const onChangeEditInput = (key: string, value: string | number) => {
+    if (typeof editedDailyReport === 'object') {
+      setEditedDailyReport({ ...editedDailyReport, [key]: value })
+    }
   }
-  const updatedDailyReports = (updatedDailyReport) => {
+  const updatedDailyReports = (updatedDailyReport: DailyReportType) => {
     const updatedDailyReportArray = dailyReports.map((dailyReport) => {
       if (dailyReport.id === updatedDailyReport.id) {
         return updatedDailyReport
@@ -76,42 +88,66 @@ const dailyReport = () => {
     setDailyReports(sortedByDate(updatedDailyReportArray))
   }
 
-  const onEditReport = (dailyReport) => {
-    editDailyReport(userId, dailyReport, editReportDateOn).then(
-      (dailyReportsData) => {
-        updatedDailyReports(dailyReportsData)
-      }
-    )
-    setEditedDailyReport(false)
+  const onEditReport = (dailyReport: DailyReportType) => {
+    if (typeof userId === 'number') {
+      editDailyReport(userId, dailyReport, editReportDateOn).then(
+        (dailyReportsData) => {
+          updatedDailyReports(dailyReportsData)
+        }
+      )
+      setEditedDailyReport({ id: NaN, reportDateOn: '', text: '', time: NaN })
+    }
   }
 
-  const sortedByDate = (data) =>
+  const sortedByDate = (data: DailyReportType[]) =>
     data
       .slice()
-      .sort((a, b) => new Date(b.reportDateOn) - new Date(a.reportDateOn))
+      .sort(
+        (a: DailyReportType, b: DailyReportType) =>
+          new Date(b.reportDateOn).getTime() -
+          new Date(a.reportDateOn).getTime()
+      )
 
-  const handleSubmit = (event) => {
+  const handleSubmit = (event: any) => {
     event.preventDefault()
-    postDailyReport(userId, inputData, reportDateOn).then((dailyReportData) => {
-      setDailyReports(sortedByDate([...dailyReports, dailyReportData]))
-    })
-    setInputData({
-      text: '',
-      time: '',
-    })
-  }
-  const onDeleteReport = (reportId) => {
-    const deleteAndGet = async () => {
-      await deleteDailyReport(userId, reportId)
-      getDailyReportsList(userId).then((dailyReportsData) => {
-        setDailyReports(dailyReportsData)
+    if (typeof userId === 'number') {
+      postDailyReport(userId, inputData, reportDateOn).then(
+        (dailyReportData) => {
+          setDailyReports(sortedByDate([...dailyReports, dailyReportData]))
+        }
+      )
+
+      setInputData({
+        text: '',
+        time: '',
       })
+    }
+  }
+  const onDeleteReport = (reportId: number) => {
+    const deleteAndGet = async () => {
+      if (typeof userId === 'number') {
+        await deleteDailyReport(userId, reportId)
+        getDailyReportsList(userId).then((dailyReportsData) => {
+          setDailyReports(dailyReportsData)
+        })
+      }
     }
     deleteAndGet()
   }
-  const onEditReportInput = (dailyReport) => {
+  const onEditReportInput = (dailyReport: DailyReportType) => {
     setEditedDailyReport(dailyReport)
     setEditReportDateOn(new Date(dailyReport.reportDateOn))
+  }
+
+  const onChangeReportDateOn = (selectedDate: Date | null) => {
+    if (selectedDate instanceof Date) {
+      setReportDateOn(selectedDate)
+    }
+  }
+  const onChangeEditReportDateOn = (selectedDate: Date | null) => {
+    if (selectedDate instanceof Date) {
+      setEditReportDateOn(selectedDate)
+    }
   }
 
   return (
@@ -129,15 +165,14 @@ const dailyReport = () => {
             editReportDateOn={editReportDateOn}
             handleSubmit={handleSubmit}
             inputData={inputData}
-            loading={loading}
             onChangeEditInput={onChangeEditInput}
             onChangeInputData={onChangeInputData}
             onDeleteReport={onDeleteReport}
             onEditReport={onEditReport}
             onEditReportInput={onEditReportInput}
             reportDateOn={reportDateOn}
-            setEditReportDateOn={setEditReportDateOn}
-            setReportDateOn={setReportDateOn}
+            onChangeEditReportDateOn={onChangeEditReportDateOn}
+            onChangeReportDateOn={onChangeReportDateOn}
             Today={Today}
           />
         )}
